Implemented full hackathon adding process.
Checkpoint 1 - the site can now deal with facebook events, adding the
details and all users into the database. When users log in for the first
time, the extra details from extended facebook permissions are
non-destructively updated on their record.

Validation is in place to prevent duplicate addition of events, both
client-side and server-side. Schedule creation is all okie, seems pretty
easy to me but need to get an outside view.

Next step is to implement a basic hackathon view, along with a schedule
widget. From there, buddy systems and pitching ideas!