Dumped current data into seeds file
Note for the current version, where for efficiency the site
only loads in the users from the facebook event a new user being
created, there's a slight bug where...

  1. If the event is currently active inside the site
  2. and your user is the first to log in

then the result is that Attendees shows as 1/1.
Seeing as you need to be a user who has not logged in, along
with the ONLY user who is registered with the site, then I believe
this shouldn't cause a serious issue. Requires you to have not
logged in and have created the hackathon at the same time, something
that is pretty paradoxical. May want to use the timestamps on the rails
server to make this go away! #2